Change parsing of number input to parseFloat().

+++ templates/admin/index_m.html.ts
@@ -87,7 +87,7 @@ ${useReact ? "" :
 				if ($this.attr('type') === 'checkbox') {
 					obj[$this.attr('id')] = $this.prop('checked');
 				} else if ($this.attr('type') === 'number') {
-					obj[$this.attr('id')] = parseInt($this.val());
+					obj[$this.attr('id')] = parseFloat($this.val());
 				} else {
 					obj[$this.attr('id')] = $this.val();
 				}
